submit F.cpp to 'F - path pass i' (abc163) [C++ (GCC 9.2.1)]

+++ 2020/0623_ABC163/F.cpp
@@ -317,7 +317,7 @@ vector<vector<Edge>> ReadTree(int N, GraphType type = GraphType::Undirected, boo
 
 struct SubTree
 {
-  int id, size;
+  ll id, size;
 };
 
 class Solve
